% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{convertToSF}
\alias{convertToSF}
\title{convertToSF:
This function converts LiDAR point cloud data to a simple features (SF) data frame
using the lidR::as.spatial and st_as_sf functions}
\usage{
convertToSF(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A simple features (SF) data frame
}
\description{
convertToSF:
This function converts LiDAR point cloud data to a simple features (SF) data frame
using the lidR::as.spatial and st_as_sf functions
}
\examples{
sf_data <- convertToSF(myLiDARData)
}
