% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{createCHM}
\alias{createCHM}
\title{createCHM:
This function creates a canopy height model out of lidar data using the rasterize_canopy function
from the "lidR" package and the p2r() algorithm}
\usage{
createCHM(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A RasterLayer object containing the canopy height model
}
\description{
createCHM:
This function creates a canopy height model out of lidar data using the rasterize_canopy function
from the "lidR" package and the p2r() algorithm
}
\examples{
chm <- createCHM(myLiDARData)
}
