% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{createDSM}
\alias{createDSM}
\title{createDSM:
This function calculates a digital surface model out of lidar data using the grid_canopy function
from the "lidR" - Package and the pitfree() algorithm}
\usage{
createDSM(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A RasterLayer Object containing the digital surface model
}
\description{
createDSM:
This function calculates a digital surface model out of lidar data using the grid_canopy function
from the "lidR" - Package and the pitfree() algorithm
}
\examples{
dsm <- createDSM(myLiDARData)
}
