% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{createDTM}
\alias{createDTM}
\title{createDTM:
This function calculates a digital terrain model out of lidar data using the grid_terrain function
from the "lidR" package and the tin() algorithm}
\usage{
createDTM(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A RasterLayer object containing the digital terrain model
}
\description{
createDTM:
This function calculates a digital terrain model out of lidar data using the grid_terrain function
from the "lidR" package and the tin() algorithm
}
\examples{
dtm <- createDTM(myLiDARData)
}
