% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{fillRaster}
\alias{fillRaster}
\title{fillRaster:
This function fills a rasterized version of LiDAR data with actual data using the rasterize function
from the "raster" package}
\usage{
fillRaster(data, raster)
}
\arguments{
\item{data}{LiDAR point cloud data}

\item{raster}{RasterLayer object}
}
\value{
A RasterLayer object with filled data
}
\description{
fillRaster:
This function fills a rasterized version of LiDAR data with actual data using the rasterize function
from the "raster" package
}
\examples{
filled_raster <- fillRaster(myLiDARData, myRaster)
}
