% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{identify_trees}
\alias{identify_trees}
\title{identify_trees:
This function identifies trees in the LiDAR point cloud data using canopy height model and tree segmentation}
\usage{
identify_trees(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A list containing tree points and segmentation information
}
\description{
identify_trees:
This function identifies trees in the LiDAR point cloud data using canopy height model and tree segmentation
}
\examples{
tree_info <- identify_trees(myLiDARData)
}
