% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{normalize}
\alias{normalize}
\title{normalize:
This function normalizes the height of lidar data using the normalize_height function
from the "lidR" package and a digital terrain model and classifies ground and noise points}
\usage{
normalize(data, dtm)
}
\arguments{
\item{data}{LiDAR point cloud data}

\item{dtm}{Digital terrain model}
}
\value{
A normalized LiDAR point cloud data object
}
\description{
normalize:
This function normalizes the height of lidar data using the normalize_height function
from the "lidR" package and a digital terrain model and classifies ground and noise points
}
\examples{
normalized_data <- normalize(myLiDARData, dtm)
}
