% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{remove_ground}
\alias{remove_ground}
\title{remove_ground:
This function removes ground points from the LiDAR point cloud data}
\usage{
remove_ground(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
LiDAR point cloud data without ground points
}
\description{
remove_ground:
This function removes ground points from the LiDAR point cloud data
}
\examples{
no_ground_data <- remove_ground(myLiDARData)
}
