% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{remove_trees}
\alias{remove_trees}
\title{remove_trees:
This function removes identified trees from the LiDAR point cloud data and removes them
and performs noise removal for optimal results}
\usage{
remove_trees(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
LiDAR point cloud data without trees
}
\description{
remove_trees:
This function removes identified trees from the LiDAR point cloud data and removes them
and performs noise removal for optimal results
}
\examples{
no_trees_data <- remove_trees(myLiDARData)
}
