% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{tree_info}
\alias{tree_info}
\title{tree_info:
This function provides information about trees in the LiDAR point cloud data, including tree diameters,
tree shapes, tree intensities, and tree heights}
\usage{
tree_info(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A data frame containing information about trees
}
\description{
tree_info:
This function provides information about trees in the LiDAR point cloud data, including tree diameters,
tree shapes, tree intensities, and tree heights
}
\examples{
tree_information <- tree_info(myLiDARData)
}
