% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{veg_cluster}
\alias{veg_cluster}
\title{veg_cluster:
This function performs k-means clustering on low vegetation points in the LiDAR point cloud data
based on height and intensity attributes}
\usage{
veg_cluster(data)
}
\arguments{
\item{data}{LiDAR point cloud data}
}
\value{
A data frame containing low vegetation points with cluster labels
}
\description{
veg_cluster:
This function performs k-means clustering on low vegetation points in the LiDAR point cloud data
based on height and intensity attributes
}
\examples{
clustered_vegetation <- veg_cluster(myLiDARData)
}
