% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestSuccession.R
\name{volume}
\alias{volume}
\title{volume:
This function calculates the volume of the LiDAR point cloud data based on a height threshold}
\usage{
volume(data, height_threshold = 5)
}
\arguments{
\item{data}{LiDAR point cloud data}

\item{height_threshold}{Height threshold for calculating volume}
}
\value{
A list containing a RasterLayer object and the total volume
}
\description{
volume:
This function calculates the volume of the LiDAR point cloud data based on a height threshold
}
\examples{
volume_data <- volume(myLiDARData, height_threshold = 5)
}
